YG-131 📂 libs 폴더 구조 변경
