+++ apis/auth/oauthApi.ts
@@ -1,6 +1,6 @@
 const POST_API_URL = "/auth"
-// const redirect_uri = "http://localhost:3000/auth"
-const redirect_uri = "https://yeogi-client.vercel.app/auth"
+const redirect_uri = "http://localhost:3000/auth"
+// const redirect_uri = "https://yeogi-client.vercel.app/auth"
 
 /**
  * @function
