+++ apis/commentApi.ts
@@ -1,7 +1,6 @@
-import { postCommentResponse } from "@/hook/type"
-
 import { postCommentProps, getCommentProps, commentIdProps, deleteCommentProps, putCommentProps } from "./type"
 import { fetchFormAPI, fetchFormAPINotToken } from "./api.utils"
+import { postCommentResponse } from "@/libs/reactQuery/type"
 
 const API_URL = "/comments"
