+++ app/(afterLogin)/detailPost/[postId]/page.tsx
@@ -13,8 +13,8 @@ import { Comment } from "./_components/comment/type"
 import { defaultIcons, handlePostIcons } from "@/constants/floatingBarIcons"
 import { useEffect } from "react"
 import DeleteModal from "@/components/commons/deleteModal"
-import { useCommentIdStore } from "@/libs/commentStore"
-import useModalStore from "@/libs/modalStore"
+import { useModalStore } from "@/libs/zustand/modal"
+import { useCommentIdStore } from "@/libs/zustand/comment"
 import LikeToComment from "./_components/comment/_components/likeToComment"
 import { Post } from "@/types/post"
 import { usePostDataStore } from "@/libs/zustand/post"
