--- libs/reactQuery/post.type.ts
@@ -1,6 +1,0 @@
-import { UpdatePost } from "@/types/post"
-
-export type updateFreeProps = {
-    postId: number
-    editedFields: UpdatePost
-}
