+++ libs/reactQuery/type.ts
@@ -1,3 +1,12 @@
+import { UpdatePost } from "@/types/post"
+
+/* post mutation type */
+export type updateFreeProps = {
+    postId: number
+    editedFields: UpdatePost
+}
+
+/* comment mutation type */
 export type postCommentResponse = {
     id: number
     content: string
@@ -7,18 +16,15 @@ export type postCommentResponse = {
     likeCount: number
     postId: number
 }
-
 export type postCommentRequest = {
     content: string
     postId: number
 }
-
 export type putCommentResponse = {
     commentId: number
     content: string
     postId: number
 }
-
 export type putCommentRequest = {
     commentId: number
     content: string
