+++ libs/zustand/modal.ts
@@ -1,12 +1,17 @@
 import { create } from "zustand"
-import { ModalStore } from "./type"
 
-const useModalStore = create<ModalStore>(set => ({
+export type ModalStore = {
+    showLoginModal: boolean
+    openLoginModal: () => void
+    closeModal: () => void
+    isDelete: boolean
+    setIsDelete: (isDelte: boolean) => void
+}
+
+export const useModalStore = create<ModalStore>(set => ({
     showLoginModal: false,
     openLoginModal: () => set({ showLoginModal: true }),
     closeModal: () => set({ showLoginModal: false }),
     isDelete: false,
     setIsDelete: (isDelete: boolean) => set({ isDelete }),
 }))
-
-export default useModalStore
